Seperate classes and interfactes Fixes #1204
