+++ packages/mcos-shared/src/types/NPS_LOBBY_COMMANDS.ts
@@ -1,0 +1,14 @@
+import { NPS_LOBBYCLIENT_COMMANDS } from './NPS_LOBBYCLIENT_COMMANDS.js';
+import { NPS_LOBBYSERVER_COMMANDS } from './NPS_LOBBYSERVER_COMMANDS.js';
+import { NpsCommandMap } from './types.js';
+
+/**
+ * @export
+ * @readonly
+ * @type {NpsCommandMap[]}
+ */
+
+export const NPS_LOBBY_COMMANDS: NpsCommandMap[] = [
+  ...NPS_LOBBYCLIENT_COMMANDS,
+  ...NPS_LOBBYSERVER_COMMANDS
+];
