+++ packages/mcos-shared/src/types/NPS_LOGIN_COMMANDS.ts
@@ -1,0 +1,10 @@
+import { NPS_LOGINCLIENT_COMMANDS } from './NPS_LOGINCLIENT_COMMANDS.js';
+import { NpsCommandMap } from './types.js';
+
+/**
+ * @export
+ * @readonly
+ * @type {NpsCommandMap[]}
+ */
+
+export const NPS_LOGIN_COMMANDS: NpsCommandMap[] = [...NPS_LOGINCLIENT_COMMANDS];
