+++ packages/mcos-shared/src/types/premadeLogin.ts
@@ -1,0 +1,268 @@
+/**
+ * This is the response packet sent on the login port in response to a UserLogin
+ *
+ * @return {Buffer}
+ */
+
+export function premadeLogin(): Buffer {
+  // TODO: #1181 Generate a dynamic login response message
+  return Buffer.from([
+    // Live Packet
+    0x06,
+    0x02,
+    0x01,
+    0x01,
+    0x21,
+    0xf9,
+    0x17,
+    0xf2,
+    0x28,
+    0x85,
+    0xd1,
+    0x47,
+    0xab,
+    0x01,
+    0x00,
+    0x00,
+    0xec,
+    0xf7,
+    0xba,
+    0x7f,
+    0x45,
+    0x62,
+    0x53,
+    0x62,
+    0xfe,
+    0x53,
+    0x7b,
+    0x03,
+    0x11,
+    0x27,
+    0x72,
+    0xbd,
+    0xa3,
+    0x3d,
+    0xa3,
+    0x06,
+    0x52,
+    0x3a,
+    0xfb,
+    0x7c,
+    0xd6,
+    0xd5,
+    0xdb,
+    0x85,
+    0x3d,
+    0x73,
+    0x66,
+    0x8f,
+    0x26,
+    0x69,
+    0x65,
+    0x07,
+    0x37,
+    0x7a,
+    0xe8,
+    0xc9,
+    0x45,
+    0x99,
+    0x6a,
+    0xaf,
+    0xe5,
+    0xdf,
+    0x1c,
+    0xbd,
+    0x1f,
+    0x30,
+    0xdc,
+    0x5a,
+    0x1a,
+    0x29,
+    0x4d,
+    0xab,
+    0x3d,
+    0x0b,
+    0x15,
+    0xdf,
+    0x33,
+    0x32,
+    0xdc,
+    0x1e,
+    0xe8,
+    0x75,
+    0x8b,
+    0x54,
+    0x34,
+    0x26,
+    0x0d,
+    0x3a,
+    0xa2,
+    0xcf,
+    0x2d,
+    0x26,
+    0x3d,
+    0x7d,
+    0xf7,
+    0xec,
+    0x3c,
+    0x52,
+    0xb2,
+    0x34,
+    0x57,
+    0xc1,
+    0x07,
+    0xd7,
+    0x6a,
+    0xd4,
+    0xdc,
+    0x1e,
+    0xd0,
+    0x07,
+    0x31,
+    0xdd,
+    0xe7,
+    0x92,
+    0x4b,
+    0xf2,
+    0x56,
+    0xc8,
+    0xb1,
+    0x00,
+    0x4d,
+    0xd6,
+    0xe8,
+    0x79,
+    0x14,
+    0xf0,
+    0x72,
+    0x71,
+    0x41,
+    0x6d,
+    0xce,
+    0x11,
+    0xe2,
+    0xae,
+    0x9d,
+    0xec,
+    0x55,
+    0x6e,
+    0xdd,
+    0xdf,
+    0xfa,
+    0xdb,
+    0x09,
+    0x74,
+    0x61,
+    0x8c,
+    0x67,
+    0xf6,
+    0xf7,
+    0x65,
+    0xf3,
+    0x98,
+    0xfd,
+    0x6d,
+    0x97,
+    0x4b,
+    0x3f,
+    0x54,
+    0x85,
+    0x4f,
+    0x2a,
+    0x69,
+    0x02,
+    0xbe,
+    0xb6,
+    0xd4,
+    0xa3,
+    0x2f,
+    0x5b,
+    0x86,
+    0x58,
+    0x19,
+    0xbd,
+    0xa9,
+    0x9e,
+    0x21,
+    0x63,
+    0x50,
+    0x9c,
+    0x43,
+    0x83,
+    0x42,
+    0xfa,
+    0xa5,
+    0x81,
+    0x5c,
+    0x1f,
+    0xad,
+    0x06,
+    0x64,
+    0x64,
+    0x7f,
+    0xe1,
+    0x2b,
+    0xdb,
+    0xd0,
+    0xee,
+    0xa6,
+    0x04,
+    0x11,
+    0x9a,
+    0x00,
+    0x00,
+    0x00,
+    0x00,
+    0xb4,
+    0x43,
+    0xf5,
+    0x00,
+    0x00,
+    0x27,
+    0x1f,
+    0x07,
+    0xa3,
+    0xbf,
+    0x17,
+    0x3d,
+    0x9b,
+    0x2e,
+    0xde,
+    0xae,
+    0xdf,
+    0x46,
+    0x2f,
+    0x30,
+    0x11,
+    0x70,
+    0xbe,
+    0x5f,
+    0x90,
+    0x46,
+    0x0c,
+    0x28,
+    0x33,
+    0xf0,
+    0x08,
+    0x88,
+    0x03,
+    0x05,
+    0xbf,
+    0xe5,
+    0x53,
+    0xcd,
+    0xfa,
+    0x45,
+    0x77,
+    0x2e,
+    0x90,
+    0xf3,
+    0x4d,
+    0xd1,
+    0x44,
+    0x89,
+    0x00,
+    0x00,
+    0x00 // (0xa6 was old val)
+  ]);
+}
